podman: exec() always closes connections
It seems that we cannot just reuse an upgraded connection
once we did the upgrade.
So, the simplest (and not-so-ugly) way out is to force a connection
close after each exec() and reopen when the next call has to be
performed.

We need to be smart enough and NOT do this trick if we never upgrade
a connection: in this case we can -and should- safely reuse it multiple
times.

Signed-off-by: Francesco Romani <fromani@redhat.com>