podman: re-implement ExecContainer
ExecContainer() has some perks, due to the fact that it
wants to use a so-called "upgraded connection".

First and foremost, the unupgraded calls (seems to) work
in a simple/request/response way.

It is not the case for the upgraded connection calls.
The server first respond as usual with a reply which tells the client
if the initial stage (the exec setup) succeeded or not.

After that, the two sides start exchanging messages.
This means that the client needs, at very least, access to a reader
to consume the messages from the server.
Hance, we added such a feature in the go varlink bindings, and
we reimplemented ExecContainer using code from libpod to
consume and dispatch these messages to stdout/stderr.

Please note that I use conditionals above because the docs are scarce,
so all of the above comes from reading the code and observing the
behaviour of the binaries.

Signed-off-by: Francesco Romani <fromani@redhat.com>