podman: consume upgraded conn messages
ExecConnection() returns first a simple reply telling
if the setup succeeded or not, and then the real, application layer-like
communication takes place.

To let the consumer consume the std{out,err}, we need to hook
the server to a dispatcher, borrowed from libpod.
We use the reimplemented call which returns a reader, from which
we extract the messages.

Signed-off-by: Francesco Romani <fromani@redhat.com>