remove terminal connection from ssh command
This is needed, because when KUBEVIRT_PROVIDER is set to oc-3.11.0 and terminal is available,
then it requires upgraded connection to attach to container. If call is changed to upgraded, then
ssh gets stuck on iopodman.GetAttachSockets().Call function and will never end.
But if communication is done over exec, then everything works as expected.

Signed-off-by: Karel Šimon <ksimon@redhat.com>