cmd: opts to pull all images needed for cluster
In this change we reorganize the way we pull the images needed
for a cluster to run.
Previously, we incrementally downloaded images as we go, alternating
download and setup. While images are cached on the host, this made
the first run less predictable, because download time (and timouts)
were intermixed with setup time (and timeouts!)
This is especially important for CI/functests, where hosts are ephemeral
VMs, so caching benefits are reduced greatly, or inexistent.

With this patch, we can move all the downloads at the beginning of
the run flow. This way, when the phase is done, we know that every
delay or timeout is caused by setup only, and not by, say, slow network.

The `run` and `pull` subcommand gained options to download all the
images needed by cluster and, in case of `run`, just exit after
the download is done. This helps warming up the image cache on the host,
so every run, including the first one, wastes no time in download.

Again, this is expected to be used and useful on CI and functests.

Being a reshuffle of steps we need anyway, we expect no real change
in behaviour for the `run` command.

Fixes: https://github.com/fromanirh/pack8s/issues/32
Signed-off-by: Francesco Romani <fromani@redhat.com>