functests: add helper runner script
For pack8s phase 1, a "functest" against a given provider is
1. bring up a cluster using a given provider
2. check the cluster is functional
3. bring down the cluster

Add scripts to allow that. So far we do #1 and #2 and we assume #3
is correct. Further patches will ensure #3 is actually done.

Worth mentioning that #2 is pretty simplistic: we just check
that API server is reported running, which is surely needed,
but maybe not enough to fully assess the cluster health.

Signed-off-by: Francesco Romani <fromani@redhat.com>